Implementar algoritmo de filosofos pensantes em pthread, além de algumas mudanças de formatação

+++ praticas/ex3/array_mult_pthread_avx.c
@@ -18,7 +18,7 @@
 
 struct thread_data {
    long thread_id;
-   int offset;
+   long unsigned int offset;
    int size;
    int last;
    float * arrayEvens;
